+++ Src/Library/changelog.md
@@ -5,4 +5,5 @@
 - 
 
 ### FIXES
-- `EndpointWithMapping<TRequest, TResponse, TEntity>` map methods becoming protected in v5.8
+- double response issue if pre-processor sent response and validator also has failures. [#info](https://discord.com/channels/933662816458645504/1080609437879914506)
+- `EndpointWithMapping<TRequest, TResponse, TEntity>` map methods becoming protected in v5.8 [#info](https://discord.com/channels/933662816458645504/1082207914376319026)